api_service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic==1.10.8
python-jose[cryptography]
httpx
python-dotenv
pytest
pytest-asyncio
redis>=5.0.0
sse-starlette>=2.0.0,<3.0.0
cachetools>=5.0.0
google-auth
google-auth-oauthlib
bcrypt
slowapi
stripe
